Development/storage/requirements.txt
confluent-kafka==2.0.2
mysql-connector-python==8.0.32
protobuf==3.20.3
